requirements.txt
tensorflow-gpu
numpy
pandas
scikit-learn
typer
imbalanced-learn